--- veles/secrets/http/testdata/src/aspnet.cs
@@ -1,20 +1,0 @@
-app.UseRouting();
-
-app.UseAuthorization();
-
-var antiforgery = app.Services.GetRequiredService<IAntiforgery>();
-
-app.Use((context, next) =>
-{
-    var requestPath = context.Request.Path.Value;
-
-    if (string.Equals(requestPath, "/", StringComparison.OrdinalIgnoreCase)
-        || string.Equals(requestPath, "/index.html", StringComparison.OrdinalIgnoreCase))
-    {
-        var tokenSet = antiforgery.GetAndStoreTokens(context);
-        context.Response.Cookies.Append("XSRF-TOKEN", tokenSet.RequestToken!,
-            new CookieOptions { HttpOnly = false });
-    }
-
-    return next(context);
-});
